requirements/base.txt
Django==5.0.3
psycopg2-binary==2.9.9
python-environ==0.4.54
django-ckeditor==6.7.1
django-mptt==0.16.0